refactor: shared telemetry + segment + thread
