+++ electron/app/createWindow.ts
@@ -5,8 +5,6 @@ import {machineIdSync} from 'node-machine-id';
 import Nucleus from 'nucleus-nodejs';
 import * as path from 'path';
 
-import {disableSegment, enableSegment, getSegmentClient} from '@utils/segment';
-
 import {APP_MIN_HEIGHT, APP_MIN_WIDTH, NEW_VERSION_CHECK_INTERVAL} from '@monokle-desktop/shared/constants/app';
 import {DEFAULT_PLUGINS} from '@monokle-desktop/shared/constants/plugin';
 import {DISABLED_TELEMETRY} from '@monokle-desktop/shared/constants/telemetry';
@@ -14,6 +12,7 @@ import {DEFAULT_TEMPLATES_PLUGIN_URL, DEPENDENCIES_HELP_URL} from '@monokle-desk
 import {AlertEnum, NewVersionCode, StartupFlags} from '@monokle-desktop/shared/models';
 import type {AlertType} from '@monokle-desktop/shared/models';
 import utilsElectronStore from '@monokle-desktop/shared/utils/electronStore';
+import {disableSegment, enableSegment, getSegmentClient} from '@monokle-desktop/shared/utils/segment';
 import {activeProjectSelector, unsavedResourcesSelector} from '@monokle-desktop/shared/utils/selectors';
 import * as Splashscreen from '@trodi/electron-splashscreen';
