+++ shared/package.json
@@ -9,6 +9,7 @@
     "dev": "nodemon -e ts -w ./src -x npm run build"
   },
   "devDependencies": {
+    "@types/analytics-node": "3.1.9",
     "@types/node": "18.0.0",
     "rimraf": "3.0.2",
     "typescript": "4.8.4"
@@ -21,10 +22,12 @@
     "@monokle/validation": "0.7.1",
     "@open-policy-agent/opa-wasm": "1.8.0",
     "@reduxjs/toolkit": "1.9.0",
+    "analytics-node": "6.2.0",
     "electron": "21.2.3",
     "electron-log": "4.4.8",
-    "electron-store": "^8.1.0",
+    "electron-store": "8.1.0",
     "lodash": "4.17.21",
+    "loglevel": "1.8.1",
     "nodemon": "2.0.20",
     "react": "18.2.0",
     "react-monaco-editor": "0.50.1",
