+++ shared/src/utils/index.ts
@@ -4,6 +4,9 @@ export * from './fileSystem';
 export * from './hotkey';
 export * from './logs';
 export * from './resource';
+export * from './segment';
 export * from './selectionHistory';
 export * from './selectors';
 export * from './shell';
+export * from './telemetry';
+export * from './thread';
