+++ src/components/molecules/FormEditor/FormEditor.tsx
@@ -22,11 +22,11 @@ import {updateResource} from '@redux/thunks/updateResource';
 
 import {ErrorPage} from '@components/organisms/ErrorPage/ErrorPage';
 
-import {trackEvent} from '@utils/telemetry';
 import {parseYamlDocument} from '@utils/yaml';
 
 import {CHANGES_BY_FORM_EDITOR} from '@monokle-desktop/shared/constants/telemetry';
 import {isInPreviewModeSelector} from '@monokle-desktop/shared/utils/selectors';
+import {trackEvent} from '@monokle-desktop/shared/utils/telemetry';
 
 import {FormArrayFieldTemplate} from './FormArrayFieldTemplate';
 import * as S from './FormEditor.styled';
