+++ src/components/organisms/ValidationPane/ValidationOverview/ValidationCard.tsx
@@ -9,9 +9,8 @@ import {updateValidationIntegration} from '@redux/reducers/main';
 import {pluginEnabledSelector} from '@redux/validation/validation.selectors';
 import {toggleValidation} from '@redux/validation/validation.slice';
 
-import {trackEvent} from '@utils/telemetry';
-
 import {ValidationIntegration} from '@monokle-desktop/shared/models';
+import {trackEvent} from '@monokle-desktop/shared/utils/telemetry';
 
 import * as S from './ValidationCard.styled';
