+++ src/utils/commands/execute.ts
@@ -5,9 +5,9 @@ import log from 'loglevel';
 import {ERROR_MSG_FALLBACK} from '@constants/constants';
 
 import {isDefined} from '@utils/filter';
-import {ensureRendererThread} from '@utils/thread';
 
 import {CommandOptions, CommandResult} from '@monokle-desktop/shared/models';
+import {ensureRendererThread} from '@monokle-desktop/shared/utils/thread';
 
 export function runCommandInMainThread(options: CommandOptions): Promise<CommandResult> {
   ensureRendererThread();
